project/db/models/learner_course.py
from django.db import models
from .learner import LearnerProfile
from .course import Course
import uuid


class LearnerCourse(models.Model):
    id = models.UUIDField(unique=True, primary_key=True, default=uuid.uuid4, editable=False)
    learner_id = models.ForeignKey(LearnerProfile, on_delete=models.CASCADE)
    course_id = models.ForeignKey(Course, on_delete=models.CASCADE)
    created_at = models.DateTimeField(auto_now_add=True)
    updated_at = models.DateTimeField(auto_now=True)

    def __str__(self):
        return str(self.id)
        